Add random string to temporary directory in solweig_algorithm.py
This commit adds a random string to the temporary directory `self.temp_dir` in solweig_algorithm.py. The goal is to make this directory unique. This matters when parallelizing solweig in a HPC environment. In this context multiple instances of solweig might run on the same compute node all trying to use the same temporary directory, which causes some instances to fail.